feat: moved token components to react-lib
