+++ src/pages/dashboard/TokenBalances.tsx
@@ -1,36 +1,24 @@
-import { Token } from '@reef-chain/react-lib';
+import { Token,Components } from '@reef-chain/react-lib';
 import Uik from '@reef-chain/ui-kit';
 import React, { useContext } from 'react';
 import { faPlus } from '@fortawesome/free-solid-svg-icons';
 import { Link } from 'react-router-dom';
 import BigNumber from 'bignumber.js';
 import TokenPricesContext from '../../context/TokenPricesContext';
-import TokenCard from './TokenCard';
+// import TokenCard from './TokenCard';
 import { BUY_URL, CREATE_ERC20_TOKEN_URL } from '../../urls';
 import { localizedStrings } from '../../l10n/l10n';
 import './loading-animation.css';
 import ReefSigners from '../../context/ReefSigners';
-import { isReefswapUI } from '../../environment';
+import { isReefswapUI, useDexConfig } from '../../environment';
+import HideBalance from '../../context/HideBalance';
+import PoolContext from '../../context/PoolContext';
 
 interface TokenBalances {
     tokens: Token[];
 }
 
-export const Skeleton = (): JSX.Element => (
-  <div className="token-card token-card--skeleton">
-    <div className="token-card__wrapper">
-      <div className="token-card__main">
-        <div className="token-card__image loading-animation" />
-        <div className="token-card__info">
-          <div className="token-card__token-info">
-            <span className="token-card__token-name loading-animation" />
-          </div>
-          <div className="token-card__token-price loading-animation" />
-        </div>
-      </div>
-    </div>
-  </div>
-);
+const {Skeleton,TokenCard} = Components;
 
 const CreateTokenButton = (): JSX.Element => (
   <Link
@@ -50,7 +38,10 @@ const balanceValue = (token: Token, price = 0): number => (new BigNumber(token.b
 
 export const TokenBalances = ({ tokens }: TokenBalances): JSX.Element => {
   const tokenPrices = useContext(TokenPricesContext);
-  const { selectedSigner, network } = useContext(ReefSigners);
+  const { selectedSigner, network,provider ,accounts } = useContext(ReefSigners);
+  const hideBalance = useContext(HideBalance);
+  const pools = useContext(PoolContext);
+  const { selectedSigner: signer, network: nw } = useContext(ReefSigners);
 
   const isReefBalanceZero = selectedSigner?.balance._hex === '0x00';
 
@@ -80,10 +71,20 @@ export const TokenBalances = ({ tokens }: TokenBalances): JSX.Element => {
       return -1;
     })
     .map((token) => (
-      <TokenCard
-        key={token.address}
-        token={token}
-        price={tokenPrices[token.address] || 0}
+      <TokenCard 
+      accounts={accounts}
+      hideBalance={hideBalance}
+      pools={pools}
+      tokenPrices={tokenPrices}
+      signer={signer}
+      nw={nw}
+      selectedSigner={selectedSigner}
+      provider={provider}
+      useDexConfig={useDexConfig}
+      isReefswapUI={isReefswapUI}
+      price={tokenPrices[token.address] || 0}
+      token={token}
+      tokens={tokens}
       />
     ));
